refactor: move components styles overrides to dedicated folder
